third_party/WebKit/LayoutTests/fast/speechsynthesis/speech-synthesis-mock-no-crash.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<script>
    description("This tests that the mock speech synthesizer is well behaved when attempting to pause/resume without first speaking.");

    if (window.internals)
        internals.enableMockSpeechSynthesizer(window);

    shouldBeUndefined("speechSynthesis.pause()");
    shouldBeUndefined("speechSynthesis.resume()");
</script>
</body>
</html>
